ibc_core_host_cosmos
Module upgrade_proposal

Provides utilities related to chain upgrades.

Structs

Plan

Specifies information about a planned upgrade and at which height it should be performed.

UpgradeChain

Event type emitted by the host chain when an upgrade plan is executed.

UpgradeClientProposal

Event type emitted by the host chain when an upgrade plan is proposed.

UpgradeProposal

Defines a governance proposal of type Content that enables the initiation of an IBC breaking upgrade and specifies the new client state that should be utilized following the upgrade.

Constants

UPGRADE_PROPOSAL_TYPE_URL

Traits

UpgradeExecutionContext

Helper context to execute client upgrades, providing methods to schedule an upgrade and store related upgraded client and consensus states.

UpgradeValidationContext

Helper context to validate client upgrades, providing methods to retrieve an upgrade plan and related upgraded client and consensus states.

Functions

execute_upgrade_client_proposal

Executes an upgrade client proposal.

Type Aliases

UpgradedClientStateRef

UpgradedConsensusStateRef