ibc_core_host_cosmos::upgrade_proposal
Trait UpgradeExecutionContext

pub trait UpgradeExecutionContext: UpgradeValidationContext {
    // Required methods
    fn schedule_upgrade(&mut self, plan: Plan) -> Result<(), UpgradeClientError>;
    fn clear_upgrade_plan(
        &mut self,
        plan_height: u64,
    ) -> Result<(), UpgradeClientError>;
    fn store_upgraded_client_state(
        &mut self,
        upgrade_path: UpgradeClientStatePath,
        client_state: UpgradedClientStateRef<Self>,
    ) -> Result<(), UpgradeClientError>;
    fn store_upgraded_consensus_state(
        &mut self,
        upgrade_path: UpgradeConsensusStatePath,
        consensus_state: UpgradedConsensusStateRef<Self>,
    ) -> Result<(), UpgradeClientError>;
}

Helper context to execute client upgrades, providing methods to schedule an upgrade and store related upgraded client and consensus states.

Required Methods

fn schedule_upgrade(&mut self, plan: Plan) -> Result<(), UpgradeClientError>

Schedules an upgrade based on the specified plan. If there is another Plan it should be overwritten.

fn clear_upgrade_plan( &mut self, plan_height: u64, ) -> Result<(), UpgradeClientError>

Clears the upgrade plan at the specified height.

fn store_upgraded_client_state( &mut self, upgrade_path: UpgradeClientStatePath, client_state: UpgradedClientStateRef<Self>, ) -> Result<(), UpgradeClientError>

Stores the upgraded client state at the specified upgrade path.

fn store_upgraded_consensus_state( &mut self, upgrade_path: UpgradeConsensusStatePath, consensus_state: UpgradedConsensusStateRef<Self>, ) -> Result<(), UpgradeClientError>

Stores the upgraded consensus state at the specified upgrade path.