ibc_core_host_cosmos::utils
Function cosmos_adr028_escrow_address

pub fn cosmos_adr028_escrow_address(
    port_id: &PortId,
    channel_id: &ChannelId,
) -> Vec<u8>

Helper function to generate an escrow address for a given port and channel ids according to the format specified in the Cosmos SDK ADR-028