ibc_core_host_cosmos::upgrade_proposal
Function execute_upgrade_client_proposal

pub fn execute_upgrade_client_proposal<Ctx>(
    ctx: &mut Ctx,
    proposal: UpgradeProposal,
) -> Result<Event, UpgradeClientError>
where
    Ctx: UpgradeExecutionContext,
    UpgradedClientStateRef<Ctx>: From<ClientState>,

Executes an upgrade client proposal.

It clears both IBC client and consensus states if a previous plan was set. Then it will schedule an upgrade and finally set the upgraded client state in the upgrade store.